backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
Pillow==10.3.0
rembg==2.0.57
python-multipart==0.0.9
onnxruntime==1.17.3
numpy==1.26.4

